+++ tests/unit/CRUD/base/base-creating-test.js
@@ -1,8 +1,7 @@
 import Ember from 'ember';
 import QueryBuilder from 'ember-flexberry-data/query/builder';
-import executeTest from './execute-odata-CRUD-test';
 
-executeTest('creating', (store, assert) => {
+export default function baseCreatingTest(store, assert) {
   assert.expect(5);
   let done = assert.async();
 
@@ -35,10 +34,13 @@ executeTest('creating', (store, assert) => {
         });
       });
     })
-    .catch(e => console.log(e, e.message))
+    .catch((e) => {
+      console.log(e, e.message);
+      throw e;
+    })
     .finally(done);
   });
-});
+}
 
 function initTestData(store) {
   return store.createRecord('ember-flexberry-dummy-application-user', {
@@ -66,11 +68,17 @@ function initTestData(store) {
         suggestion: sug,
       }).save()
 
+      // It is necessary to fill 'detail' at 'master' in offline.
+      .then((comment) => store._isOnline() ? Ember.RSVP.resolve(comment) : sug.save().then(() => Ember.RSVP.resolve(comment)))
+
       .then((comment) =>
         store.createRecord('ember-flexberry-dummy-comment-vote', {
           applicationUser: user,
           comment: comment
         }).save()
+
+        // It is necessary to fill 'detail' at 'master' in offline.
+        .then((vote) => store._isOnline() ? Ember.RSVP.resolve(vote) : comment.save().then(() => Ember.RSVP.resolve(vote)))
 
         .then((vote) =>
           new Ember.RSVP.Promise((resolve) =>