+++ tests/unit/CRUD/odata/odata-creating-test.js
@@ -1,0 +1,6 @@
+import executeTest from './execute-odata-test';
+import baseCreatingTest from '../base/base-creating-test';
+
+executeTest('creating', (store, assert) => {
+  baseCreatingTest(store, assert);
+});