+++ tests/unit/CRUD/offline/offline-reading-complex-predicates-test.js
@@ -1,0 +1,6 @@
+import executeTest from './execute-offline-test';
+import readingComplexPredicates from '../base/base-reading-complex-predicates-test';
+
+executeTest('reading | predicates | complex predicates', (store, assert) => {
+  readingComplexPredicates(store, assert);
+});